move loadMigrations to db package src root
